Return ContainerData from container create
